Add missing build-depends, switch to debhelper-compat
We depend upon libssl-dev for FIT tool builds (and probably others), but
those tools are used only on the build machine as far as I can tell. Since
I don't know how to tell pbuilder that yet, consider this change as incomplete.

I switched to debhelper-compat on the advice of lintian.